bug fix
deleteNights.py has been updated to use the correct version of
`connectChironDB`. It was previously using the version of
`connectChironDB` that established a pyMySQL connections instead of
the SQLAlchemy engine, but did not specify the `legacy=True` optional
argument. It now specifies `legacy=True` to continue using the PyMySQL
legacy connection.